Split LedgerDB into its own file
