+++ ledger/migration/src/lib.rs
@@ -8,10 +8,14 @@
 use lmdb::{DatabaseFlags, Environment, Transaction, WriteFlags};
 use mc_common::logger::{log, Logger};
 use mc_ledger_db::{
-    key_bytes_to_u64, tx_out_store::TX_OUT_INDEX_BY_PUBLIC_KEY_DB_NAME, u64_to_key_bytes, Error,
-    LedgerDbMetadataStoreSettings, MetadataStore, MintConfigStore, MintTxStore, TxOutStore,
-    TxOutsByBlockValue, BLOCK_NUMBER_BY_TX_OUT_INDEX, COUNTS_DB_NAME, MAX_LMDB_DATABASES,
-    MAX_LMDB_FILE_SIZE, NUM_BLOCKS_KEY, TX_OUTS_BY_BLOCK_DB_NAME,
+    key_bytes_to_u64,
+    ledger_db::{
+        LedgerDbMetadataStoreSettings, TxOutsByBlockValue, BLOCK_NUMBER_BY_TX_OUT_INDEX,
+        COUNTS_DB_NAME, MAX_LMDB_DATABASES, MAX_LMDB_FILE_SIZE, NUM_BLOCKS_KEY,
+        TX_OUTS_BY_BLOCK_DB_NAME,
+    },
+    tx_out_store::TX_OUT_INDEX_BY_PUBLIC_KEY_DB_NAME,
+    u64_to_key_bytes, Error, MetadataStore, MintConfigStore, MintTxStore, TxOutStore,
 };
 use mc_util_lmdb::MetadataStoreError;
 use mc_util_serial::decode;
